[CL-1941] Move Settings to shared folder
