+++ front/app/components/ContentBuilder/messages.ts
@@ -5,4 +5,12 @@ export default defineMessages({
     id: 'app.containers.admin.ContentBuilder.urlPlaceholder',
     defaultMessage: 'https://example.com',
   },
+  delete: {
+    id: 'app.containers.admin.ContentBuilder.delete',
+    defaultMessage: 'Delete',
+  },
+  a11y_closeSettingsPanel: {
+    id: 'app.containers.AdminPage.ProjectDescription.a11y_closeSettingsPanel',
+    defaultMessage: 'Close settings panel',
+  },
 });
